+++ CHANGELOG.md
@@ -4,11 +4,18 @@
 
 Intermediate Releases: TBD
 
-#### Enhancements and Changes:
+#### Enhancements and Changes
 
+- removed the service= parameter from all providers so we can use the
+  StorageABC class that does not include it
+- reorganized all tests and work on tests 01-05
 - Added README.md and CHANGELOG.md
 - reorganized the tests
 
-#### Bug Fixes:
+#### Bug Fixes
+
+#### Known Bugs
+
+None of this has been tested :-)